+++ apps/backend-go-fiber/cmd/server/service_other.go
@@ -1,9 +1,0 @@
-//go:build !darwin
-
-package main
-
-import "fmt"
-
-func runServiceCommand(action string) error {
-	return fmt.Errorf("%s is only supported on macOS for now", action)
-}